Lighten Tensor.Factories file
